Makes the components importable separately from the plugin
