+++ babel.config.js
@@ -12,5 +12,29 @@ module.exports = {
       }
     ]
   ],
-  plugins: ['@babel/plugin-proposal-optional-chaining']
+  plugins: ['@babel/plugin-proposal-optional-chaining'],
+  env: {
+    lib: {
+      presets: [
+        [
+          '@babel/preset-env', {
+            modules: false,
+            // https://babeljs.io/docs/en/babel-preset-env#usebuiltins
+            useBuiltIns: 'usage',
+            corejs: {
+              version: 3
+            }
+          }
+        ]
+      ],
+      plugins: [
+        ['module-resolver', {
+          root: ['.'],
+          alias: {
+            '~': './lib-temp'
+          }
+        }]
+      ]
+    }
+  }
 }
