+++ package.json
@@ -5,7 +5,7 @@
   "scripts": {
     "commit": "git-cz",
     "dev": "vue-cli-service serve ./demo/main.js",
-    "build": "rollup -c --environment BUILD:production",
+    "build": "rollup -c --environment BUILD:production && node build/build-lib.js",
     "build-watch": "rollup -w -c",
     "lint": "eslint ./src --ext .ts,.js,.vue --ignore-path .gitignore",
     "lint-fix": "yarn lint --fix"
@@ -19,6 +19,7 @@
     "tiptap-extensions": "^1.14.0"
   },
   "devDependencies": {
+    "@babel/cli": "^7.6.4",
     "@babel/core": "^7.6.4",
     "@babel/plugin-proposal-optional-chaining": "^7.2.0",
     "@babel/preset-env": "^7.6.3",
@@ -35,6 +36,7 @@
     "@vue/cli-plugin-typescript": "^3.5.1",
     "@vue/cli-service": "^3.5.1",
     "@zerollup/ts-transform-paths": "^1.7.0",
+    "babel-plugin-module-resolver": "^4.0.0",
     "commitizen": "^4.0.3",
     "core-js": "3",
     "cross-env": "^6.0.3",
@@ -55,7 +57,7 @@
     "rollup-plugin-replace": "^2.2.0",
     "rollup-plugin-terser": "^4.0.4",
     "rollup-plugin-typescript2": "^0.20.1",
-    "rollup-plugin-vue": "^4.7.2",
+    "rollup-plugin-vue": "^5.1.9",
     "sass": "^1.23.0",
     "sass-loader": "^7.0.0",
     "semantic-release": "^15.13.3",
@@ -104,6 +106,7 @@
   "files": [
     "dist/*",
     "types/*",
-    "src/*"
+    "src/*",
+    "lib/*"
   ]
 }
